pyproject.toml
[project]
name = "dfresearch"
version = "0.1.0"
description = "Autonomous deepfake detection research for BitMind — baseline models and autoresearch loop"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "timm>=1.0.0",
    "transformers>=4.40.0",
    "peft>=0.11.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.23.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "scikit-image>=0.22.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "einops>=0.7.0",
    "decord>=0.6.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "pyarrow>=15.0.0",
]

[project.optional-dependencies]
wandb = ["wandb>=0.16.0"]

[project.scripts]
dfresearch = "dfresearch.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
